+++ components/ActionSheet/src/private/MDCActionSheetItemTableViewCell.m
@@ -16,7 +16,7 @@
 
 #import "MaterialTypography.h"
 
-static const CGFloat kLabelAlpha = 0.87f;
+static const CGFloat kLabelAlpha = (CGFloat)0.87;
 static const CGFloat kImageLeadingPadding = 16;
 static const CGFloat kImageTopPadding = 16;
 static const CGFloat kImageHeightAndWidth = 24;
@@ -201,7 +201,7 @@ static const CGFloat kActionItemTitleVerticalPadding = 18;
   _inkColor = inkColor;
   // If no ink color then reset to the default ink color
   self.inkTouchController.defaultInkView.inkColor =
-      inkColor ?: [[UIColor alloc] initWithWhite:0 alpha:0.14f];
+      inkColor ?: [[UIColor alloc] initWithWhite:0 alpha:(CGFloat)0.14];
 }
 
 - (void)setImageRenderingMode:(UIImageRenderingMode)imageRenderingMode {
