+++ components/Buttons/tests/unit/ButtonsColorThemerTests.m
@@ -17,7 +17,7 @@
 #import "MaterialButtons.h"
 #import "MaterialButtons+ColorThemer.h"
 
-static const CGFloat kEpsilonAccuracy = 0.001f;
+static const CGFloat kEpsilonAccuracy = (CGFloat)0.001;
 
 @interface ButtonsColorThemerTests : XCTestCase
 
@@ -40,7 +40,7 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
                         UIColor.clearColor);
   XCTAssertEqualObjects([button titleColorForState:UIControlStateNormal], colorScheme.primaryColor);
   XCTAssertEqualObjects([button titleColorForState:UIControlStateDisabled],
-                        [colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f]);
+                        [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.38]);
   NSUInteger maximumStateValue = UIControlStateNormal | UIControlStateSelected |
       UIControlStateHighlighted | UIControlStateDisabled;
   for (NSUInteger state = 0; state <= maximumStateValue; ++state) {
@@ -66,11 +66,11 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
   XCTAssertEqualObjects([button backgroundColorForState:UIControlStateNormal],
                         colorScheme.primaryColor);
   XCTAssertEqualObjects([button backgroundColorForState:UIControlStateDisabled],
-                        [colorScheme.onSurfaceColor colorWithAlphaComponent:0.12f]);
+                        [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.12]);
   XCTAssertEqualObjects([button titleColorForState:UIControlStateNormal],
                         colorScheme.onPrimaryColor);
   XCTAssertEqualObjects([button titleColorForState:UIControlStateDisabled],
-                        [colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f]);
+                        [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.38]);
   NSUInteger maximumStateValue = UIControlStateNormal | UIControlStateSelected |
       UIControlStateHighlighted | UIControlStateDisabled;
   for (NSUInteger state = 0; state <= maximumStateValue; ++state) {
@@ -115,15 +115,15 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
         XCTAssertEqual([button backgroundColorForState:state], colorScheme.primaryColor);
       }
     } else {
-      XCTAssert(
-          CGColorEqualToColor([button titleColorForState:state].CGColor,
-                              [colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f].CGColor));
-      XCTAssert(
-          CGColorEqualToColor([button backgroundColorForState:state].CGColor,
-                              [colorScheme.onSurfaceColor colorWithAlphaComponent:0.12f].CGColor));
+      XCTAssert(CGColorEqualToColor(
+          [button titleColorForState:state].CGColor,
+          [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.38].CGColor));
+      XCTAssert(CGColorEqualToColor(
+          [button backgroundColorForState:state].CGColor,
+          [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.12].CGColor));
     }
   }
-  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, 0.001f);
+  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, (CGFloat)0.001);
 }
 
 - (void)testMDCFlatButtonColorThemer {
@@ -157,15 +157,15 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
         XCTAssertEqual([button backgroundColorForState:state], UIColor.clearColor);
       }
     } else {
-      XCTAssert(
-          CGColorEqualToColor([button titleColorForState:state].CGColor,
-                              [colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f].CGColor),
-          @"state:%lu", (unsigned long)state);
+      XCTAssert(CGColorEqualToColor(
+                    [button titleColorForState:state].CGColor,
+                    [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.38].CGColor),
+                @"state:%lu", (unsigned long)state);
       XCTAssertEqual([button backgroundColorForState:state], UIColor.clearColor, @"state:%lu",
                      (unsigned long)state);
     }
   }
-  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, 0.001f);
+  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, (CGFloat)0.001);
 }
 
 - (void)testMDCRaisedButtonColorThemer {
@@ -200,15 +200,15 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
         XCTAssertEqual([button backgroundColorForState:state], colorScheme.primaryColor);
       }
     } else {
-      XCTAssert(
-          CGColorEqualToColor([button titleColorForState:state].CGColor,
-                              [colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f].CGColor));
-      XCTAssert(
-          CGColorEqualToColor([button backgroundColorForState:state].CGColor,
-                              [colorScheme.onSurfaceColor colorWithAlphaComponent:0.12f].CGColor));
+      XCTAssert(CGColorEqualToColor(
+          [button titleColorForState:state].CGColor,
+          [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.38].CGColor));
+      XCTAssert(CGColorEqualToColor(
+          [button backgroundColorForState:state].CGColor,
+          [colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.12].CGColor));
     }
   }
-  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, 0.001f);
+  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, (CGFloat)0.001);
 }
 
 - (void)testMDCFloatingButtonColorThemer {
@@ -236,7 +236,7 @@ static const CGFloat kEpsilonAccuracy = 0.001f;
     }
   }
 
-  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, 0.001f);
+  XCTAssertEqualWithAccuracy(button.disabledAlpha, 1, (CGFloat)0.001);
 }
