+++ components/FlexibleHeader/examples/FlexibleHeaderUINavigationBarExample.m
@@ -155,7 +155,7 @@ static const CGFloat kFlexibleHeaderMinHeight = 200;
   [self.view addSubview:self.fhvc.view];
   [self.fhvc didMoveToParentViewController:self];
 
-  self.fhvc.headerView.backgroundColor = [UIColor colorWithWhite:0.1f alpha:1];
+  self.fhvc.headerView.backgroundColor = [UIColor colorWithWhite:(CGFloat)0.1 alpha:1];
 }
 
 - (void)viewWillAppear:(BOOL)animated {
