+++ components/Slider/examples/SliderCollectionViewController.m
@@ -224,7 +224,7 @@ static CGFloat const kSliderVerticalMargin = 12;
 
     model = [[MDCSliderModel alloc] init];
     model.labelString = @"Default slider";
-    model.value = 0.66f;
+    model.value = (CGFloat)0.66;
     [_sliders addObject:model];
 
     model = [[MDCSliderModel alloc] init];
@@ -237,7 +237,7 @@ static CGFloat const kSliderVerticalMargin = 12;
     model = [[MDCSliderModel alloc] init];
     model.labelString = @"Discrete slider with numeric value label";
     model.numDiscreteValues = 5;
-    model.value = 0.2f;
+    model.value = (CGFloat)0.2;
     [_sliders addObject:model];
 
     model = [[MDCSliderModel alloc] init];
@@ -250,22 +250,22 @@ static CGFloat const kSliderVerticalMargin = 12;
     model = [[MDCSliderModel alloc] init];
     model.labelString = @"Dark themed slider";
     model.labelColor = [UIColor whiteColor];
-    model.trackBackgroundColor = [[UIColor whiteColor] colorWithAlphaComponent:0.3f];
+    model.trackBackgroundColor = [[UIColor whiteColor] colorWithAlphaComponent:(CGFloat)0.3];
     model.sliderColor = MDCPalette.bluePalette.tint500;
     model.bgColor = [UIColor darkGrayColor];
-    model.value = 0.2f;
+    model.value = (CGFloat)0.2;
     [_sliders addObject:model];
 
     model = [[MDCSliderModel alloc] init];
     model.labelString = @"Anchored slider";
-    model.anchorValue = 0.5f;
-    model.value = 0.7f;
+    model.anchorValue = (CGFloat)0.5;
+    model.value = (CGFloat)0.7;
     [_sliders addObject:model];
 
     model = [[MDCSliderModel alloc] init];
     model.labelString = @"Disabled slider";
-    model.value = 0.5f;
-    model.anchorValue = 0.1f;
+    model.value = (CGFloat)0.5;
+    model.anchorValue = (CGFloat)0.1;
     model.enabled = NO;
     [_sliders addObject:model];
